Conflicts:
	django/db/models/deletion.py

+++ django/contrib/gis/geos/geometry.py
@@ -115,7 +115,7 @@ class GEOSGeometry(GEOSBase, ListMixin):
         Destroys this Geometry; in other words, frees the memory used by the
         GEOS C++ object.
         """
-        if self._ptr:
+        if self._ptr and capi:
             capi.destroy_geom(self._ptr)
 
     def __copy__(self):
@@ -376,9 +376,8 @@ class GEOSGeometry(GEOSBase, ListMixin):
     @property
     def ewkt(self):
         """
-        Returns the EWKT (WKT + SRID) of the Geometry.  Note that Z values
-        are *not* included in this representation because GEOS does not yet
-        support serializing them.
+        Returns the EWKT (SRID + WKT) of the Geometry. Note that Z values
+        are only included in this representation if GEOS >= 3.3.0.
         """
         if self.get_srid():
             return 'SRID=%s;%s' % (self.srid, self.wkt)
